refactor: extract IIconComponent interface to separate file
