+++ docs/docs/auto-docs/shared-components/IconComponent/IconComponent/functions/default.md
@@ -6,13 +6,13 @@
 
 > **default**(`props`): `Element`
 
-Defined in: [src/shared-components/IconComponent/IconComponent.tsx:66](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/IconComponent/IconComponent.tsx#L66)
+Defined in: [src/shared-components/IconComponent/IconComponent.tsx:59](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/IconComponent/IconComponent.tsx#L59)
 
 ## Parameters
 
 ### props
 
-[`IIconComponent`](../interfaces/IIconComponent.md)
+[`IIconComponent`](../../../../types/shared-components/IconComponent/interface/interfaces/IIconComponent.md)
 
 ## Returns
