--- docs/docs/auto-docs/shared-components/IconComponent/IconComponent/interfaces/IIconComponent.md
@@ -1,39 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: IIconComponent
-
-Defined in: [src/shared-components/IconComponent/IconComponent.tsx:59](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/IconComponent/IconComponent.tsx#L59)
-
-## Properties
-
-### fill?
-
-> `optional` **fill**: `string`
-
-Defined in: [src/shared-components/IconComponent/IconComponent.tsx:61](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/IconComponent/IconComponent.tsx#L61)
-
-***
-
-### height?
-
-> `optional` **height**: `string`
-
-Defined in: [src/shared-components/IconComponent/IconComponent.tsx:62](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/IconComponent/IconComponent.tsx#L62)
-
-***
-
-### name
-
-> **name**: `string`
-
-Defined in: [src/shared-components/IconComponent/IconComponent.tsx:60](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/IconComponent/IconComponent.tsx#L60)
-
-***
-
-### width?
-
-> `optional` **width**: `string`
-
-Defined in: [src/shared-components/IconComponent/IconComponent.tsx:63](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/IconComponent/IconComponent.tsx#L63)
