+++ docs/docs/auto-docs/types/shared-components/IconComponent/interface/interfaces/IIconComponent.md
@@ -1,0 +1,41 @@
+[Admin Docs](/)
+
+***
+
+# Interface: IIconComponent
+
+Defined in: [src/types/shared-components/IconComponent/interface.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/IconComponent/interface.ts#L4)
+
+Props for IconComponent.
+
+## Properties
+
+### fill?
+
+> `optional` **fill**: `string`
+
+Defined in: [src/types/shared-components/IconComponent/interface.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/IconComponent/interface.ts#L6)
+
+***
+
+### height?
+
+> `optional` **height**: `string`
+
+Defined in: [src/types/shared-components/IconComponent/interface.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/IconComponent/interface.ts#L7)
+
+***
+
+### name
+
+> **name**: `string`
+
+Defined in: [src/types/shared-components/IconComponent/interface.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/IconComponent/interface.ts#L5)
+
+***
+
+### width?
+
+> `optional` **width**: `string`
+
+Defined in: [src/types/shared-components/IconComponent/interface.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/IconComponent/interface.ts#L8)
