+++ src/shared-components/IconComponent/IconComponent.tsx
@@ -53,15 +53,8 @@ import VenueIcon from 'assets/svgs/venues.svg?react';
 import RequestsIcon from 'assets/svgs/requests.svg?react';
 import ExitToAppIcon from '@mui/icons-material/ExitToApp';
 import { MdOutlineVolunteerActivism } from 'react-icons/md';
-
+import { IIconComponent } from 'types/shared-components/IconComponent/interface';
 import type { JSX } from 'react';
-
-export interface IIconComponent {
-  name: string;
-  fill?: string;
-  height?: string;
-  width?: string;
-}
 
 const iconComponent = (props: IIconComponent): JSX.Element => {
   switch (props.name) {
