+++ src/types/shared-components/IconComponent/interface.ts
@@ -1,0 +1,9 @@
+/**
+ * Props for IconComponent.
+ */
+export interface IIconComponent {
+  name: string;
+  fill?: string;
+  height?: string;
+  width?: string;
+}
